requirements.txt
boto3==1.9.156
botocore==1.12.156
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-slugger==1.1.2
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.14
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
psycopg2==2.8.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.3
